CTCP-4861: content working
