+++ app/connectors/ReferenceDataConnector.scala
@@ -94,10 +94,10 @@ class ReferenceDataConnector @Inject() (config: FrontendAppConfig, http: HttpCli
     http.GET[NonEmptySet[DocumentType]](url, headers = version2Header, queryParams = queryParams).map(_.head)
   }
 
-  def getAdditionalReferenceType(typeValue: String)(implicit ec: ExecutionContext, hc: HeaderCarrier): Future[AdditionalReference] = {
+  def getAdditionalReferenceType(typeValue: String)(implicit ec: ExecutionContext, hc: HeaderCarrier): Future[AdditionalReferenceType] = {
     val queryParams: Seq[(String, String)] = Seq("data.documentType" -> typeValue)
     val url                                = s"${config.referenceDataUrl}/lists/AdditionalReference"
-    http.GET[NonEmptySet[AdditionalReference]](url, headers = version2Header, queryParams = queryParams).map(_.head)
+    http.GET[NonEmptySet[AdditionalReferenceType]](url, headers = version2Header, queryParams = queryParams).map(_.head)
   }
 
   def getTransportDocument(typeValue: String)(implicit ec: ExecutionContext, hc: HeaderCarrier): Future[DocumentType] = {
